+++ MyQuizMobile/MyQuizMobile/MyQuizMobile/DataModel/QuestionBlock.cs
@@ -1,6 +1,6 @@
 ﻿namespace MyQuizMobile.DataModel {
     public partial class QuestionBlock {
-        public override int Id { get; set; }
+        public override long Id { get; set; }
         public string Title { get; set; }
     }
 }
